+++ .build/composer.json
@@ -4,7 +4,7 @@
     "prefer-stable": true,
     "require": {
         "php": "^7.3",
-        "narrowspark/coding-standard": "^5.0.0",
+        "narrowspark/coding-standard": "^5.1.0",
         "wikimedia/composer-merge-plugin": "^1.4.1"
     },
     "extra": {
